Extract GraphQL, remove redundant interfaces, make functional
