+++ src/components/member/tabs/index.js
@@ -1,5 +1,4 @@
 import ClaimsTab from 'components/member/tabs/ClaimsTab'
-import { MemberDebtComponent } from 'components/member/tabs/DebtTab'
 import DetailsTab from 'components/member/tabs/DetailsTab'
 import MemberFile from 'components/member/tabs/FileTab'
 import PaymentsTab from 'components/member/tabs/payments-tab'
@@ -13,6 +12,7 @@ import { Tab } from 'semantic-ui-react'
 import styled from 'react-emotion'
 import { useContractMarketInfo } from 'graphql/use-get-member-contract-market-info'
 import { AccountTab } from './account-tab'
+import { DebtTab } from 'components/member/tabs/debt-tab'
 
 const TabContainer = styled(Tab.Pane)`
   &&& {
@@ -49,7 +49,9 @@ const memberPagePanes = (props) => {
     },
     {
       menuItem: 'Claims',
-      render: () => <TabItem props={{...props, memberId}} TabContent={ClaimsTab} />,
+      render: () => (
+        <TabItem props={{ ...props, memberId }} TabContent={ClaimsTab} />
+      ),
     },
     {
       menuItem: 'Files',
@@ -102,7 +104,9 @@ const memberPagePanes = (props) => {
     },
     {
       menuItem: 'Debt',
-      render: () => <TabItem props={props} TabContent={MemberDebtComponent} />,
+      render: () => (
+        <TabItem props={{ ...props, memberId }} TabContent={DebtTab} />
+      ),
     },
   )
