+++ src/graphql/whitelist-member.graphql
@@ -1,0 +1,3 @@
+mutation whitelistMember($memberId: ID!) {
+  whitelistMember(memberId: $memberId)
+}
